api/requirements.txt
Django==3.0.7
djangorestframework==3.15.0
psycopg2-binary==2.8.6
